move grpc_service to adapter
