+++ gateway/infra/adapter/command_handler.go
@@ -1,6 +1,9 @@
 package adapter
 
-import "github.com/it-chain/it-chain-Engine/gateway/api"
+import (
+	"github.com/it-chain/it-chain-Engine/gateway"
+	"github.com/it-chain/it-chain-Engine/gateway/api"
+)
 
 type ConnectionCommandHandler struct {
 	connectionApi api.ConnectionApi
@@ -8,4 +11,8 @@ type ConnectionCommandHandler struct {
 
 type MessageCommandHandler struct {
 	messageApi api.MessageApi
+}
+
+func (c ConnectionCommandHandler) HandleGrpcDeliverCommand(command gateway.GrpcDeliverCommand) {
+
 }
